- Arrow keys rotate the camera ; 'A' toggles a debug X/Y/Z axis rendered as thin cylinders   -vec3_dot/vec3_len/vec3_scale standardised to use double throughout     - FOV must now be a whole number     - Split parse_shapes.c (sphere) and parse_shapes2.c (plane/cylinder)

+++ srcs/vec3/vec3_ops1.c
@@ -22,14 +22,14 @@ t_vec3	vec3_sub(t_vec3 a, t_vec3 b)
 	return ((t_vec3){a.x - b.x, a.y - b.y, a.z - b.z});
 }
 
-t_vec3	vec3_scale(t_vec3 v, float s)
+t_vec3	vec3_scale(t_vec3 v, double s)
 {
 	return ((t_vec3){v.x * s, v.y * s, v.z * s});
 }
 
 // Compute the dot product of two vectors
 // a . b = (a.x * b.x + a.y * b.y + a.z * b.z)
-float	vec3_dot(t_vec3 a, t_vec3 b)
+double	vec3_dot(t_vec3 a, t_vec3 b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
